+++ packages/expo-cli/src/commands/index.ts
@@ -7,16 +7,17 @@ const COMMANDS = [
   require('./auth/whoami'),
   require('./build'),
   require('./client'),
-  require('./config/config'),
   require('./credentials'),
   require('./customize'),
-  require('./diagnostics'),
-  require('./doctor'),
   require('./eject'),
   require('./expokit/bundle-assets'),
   require('./expokit/prepare-detached-build'),
   require('./export'),
   require('./fetch'),
+  require('./info/config/config'),
+  require('./info/diagnostics'),
+  require('./info/doctor'),
+  require('./info/upgrade'),
   require('./init'),
   require('./install'),
   require('./prebuild'),
@@ -25,7 +26,6 @@ const COMMANDS = [
   require('./run'),
   require('./send'),
   require('./start'),
-  require('./upgrade'),
   require('./upload'),
   require('./url'),
   require('./webhooks'),
