+++ packages/expo-cli/src/commands/info/config/config.ts
@@ -1,6 +1,6 @@
 import type { Command } from 'commander';
 
-import { applyAsyncActionProjectDir } from '../utils/applyAsyncAction';
+import { applyAsyncActionProjectDir } from '../../utils/applyAsyncAction';
 
 export default function (program: Command) {
   applyAsyncActionProjectDir(
