+++ packages/expo-cli/src/commands/info/doctor/windows.ts
